mojo1: migrate stack to Mojo 1.0.0/MAX 26.5; h3 sage scratch, adaptive step cache, audio veto
Toolchain: default pixi env is MAX 26.5 (Mojo 1.0.0 ed45d567); b1 kept as
the mojo3 env for rollback. Migration: std.gpu partial move to max.gpu
(index intrinsics stayed std.gpu), single-param enqueue_function (1,403
sites), LayoutTensor keyword ctor + origin idiom (3,392 sites), kernel
Int->Int32/Int64 widening (687 kernels, 2,078 params, launch casts read
from signatures), Float32**0.5 -> **Float32(0.5), perf_counter_ns Int,
origin-strictness and null-pointer fixes. Parity: all gates digit-identical
to the b1 anchors (sage cos 0.9999350873412836, step-cache
0.9999578397087012, scratch bit-mismatches 0) and a same-seed product run
byte-identical (latent sha256 match).

H3 speed/reliability: SageInt8Scratch preallocated per run kills the
per-call 1.6-3.2 GiB transient churn OOM (video-0177); controlled A/B at
1344x768x243 measured cuDNN 234.6 s vs Sage 155.7 s per hot evaluation
(1.51x). Step cache is adaptive (accumulated drift budget 0.24, 3-eval
exact tail; decoded A/B accepted at 1.50x, 8/19 reuse) with an audio
DECISION VETO (tighter 0.06/0.12 audio thresholds; decoded High+Sage audio
corr 0.8683->0.8809, attenuation halved to -2.5 dB); the application-side
audio mask measured harmful (corr 0.2140) and ships unwired with its
numbers in the docstring. SquareQ gains the convrot_w8_v1 offline packer
(sharded sources, resume byte-identical) with the runtime deliberately
parked by measurement on real H3 tensors (+2.78 dB mean vs per-channel
int8, mild 2.9-15.3x channel-outlier ratios).

+++ serenitymojo/models/scail2/scail2_block.mojo
@@ -6,7 +6,7 @@
 # added before the single cross_attn.o projection, exactly as the oracle does.
 
 from std.collections import Dict
-from std.gpu.host import DeviceContext
+from max.gpu.host import DeviceContext
 from std.memory import ArcPointer
 
 from serenitymojo.io.dtype import STDtype
@@ -98,7 +98,7 @@ def scail2_block_forward[
         # padding would be a different attention distribution.
         raise Error("SCAIL-2 oracle attends all 512 padded text tokens")
     var dim = cfg.dim
-    var scale = 1.0 / Float32(DH) ** 0.5
+    var scale = 1.0 / Float32(DH) ** Float32(0.5)
 
     # e0 is the batch-1 broadcastable [1,1,6,D] form. The official expanded
     # [1,6,D] row is identical for every token, so no S-sized F32 copy is made.
